+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ## 2.5.8 //
 
+### 💪 Improvements
+- Fix Service Mark, when indicated in brackets, i.e. `Brand(sm)` → `Brand℠`
+
 ### 🔨 Maintenance
 - Add a build automation
 - Automate copyright notice updates in source and dist files
+- Refactor code for Registered Trademark (®) and Trademark (™)